requirements.txt
cbor2
pyzmq
pyvis
aiohttp
pycryptodome
